Change IAnimationTimer to return double

+++ src/main/kotlin/net/barribob/boss/animation/IAnimationTimer.kt
@@ -1,5 +1,5 @@
 package net.barribob.boss.animation
 
 interface IAnimationTimer {
-    fun getCurrentTick(): Float
+    fun getCurrentTick(): Double
 }
